+++ src/pages/RentalManagement/AddSerializedAsset.tsx
@@ -8,7 +8,7 @@ import { Box, CircularProgress } from "@material-ui/core";
 import SearchBox from '../../components/Helpers/SearchBox'
 import routes from "../../components/Helpers/Routes";
 import CustomAgGrid, { reducer, intialState } from "../../components/AgGridComponents/CustomAgGrid";
-import { productInventory, isObjectEmpty, gridLoadingTimeout, CustomDialogTransition } from '../../constants/helpers';
+import { productInventory, isObjectEmpty, gridLoadingTimeout, CustomDialogTransition, getLocalStorageArrayData } from '../../constants/helpers';
 import {
     CreatedByRenderer,
     UpdatedByRenderer
@@ -18,6 +18,9 @@ import { useData } from "../../StateProvider/Provider";
 import Dialog from "@material-ui/core/Dialog/Dialog";
 import CustomDialogHeader from "../../components/CustomDialog/CustomDialogHeader";
 import CustomDialogContent from "../../components/CustomDialog/CustomDialogContent";
+
+const addSerializedAssetsRenderedFrom = "addSerializedAssets";
+const localStorageSelectedRecords = `${addSerializedAssetsRenderedFrom}_selected`;
 
 const AddSerializedAsset = ({ isAdding, addSerializedAsset, handleSerializedAssetClose, selectedProducts }) => {
     const toastConfig = useContext(CustomToastContext)
@@ -35,9 +38,10 @@ const AddSerializedAsset = ({ isAdding, addSerializedAsset, handleSerializedAsse
         fetchProductInventory()
     }, [page, limit, filters, sorting, search]);
 
-
     useEffect(() => {
         let tempProducts = serializedProducts
+        const alreadyStoredSelectedRecords = [...getLocalStorageArrayData(localStorageSelectedRecords)];
+
         if (tempProducts.length === 0) {
             selectedProducts.map(d => {
                 if (d.productName) {
@@ -64,29 +68,28 @@ const AddSerializedAsset = ({ isAdding, addSerializedAsset, handleSerializedAsse
             tempProducts = []
             selectedProducts.map(d => {
                 if (d.productName) {
-                    tempProducts.push({ "id": d._id, "name": d.productName, "qty": d?.qty - selectedRecords.filter(obj => obj.product.optionValue === d._id).length })
+                    tempProducts.push({ "id": d._id, "name": d.productName, "qty": d?.qty - alreadyStoredSelectedRecords.filter(obj => obj.product.optionValue === d._id).length })
                 }
                 if (d.packageName && d?.products?.length > 0) {
                     d.products.map(u => {
                         if (tempProducts.find(obj => obj.id === d?.productDetail?._id)) {
-                            tempProducts.find(obj => obj.id === d?.productDetail?._id).qty = u?.qty * d?.qty + tempProducts.find(obj => obj.id === d?.productDetail?._id).qty - selectedRecords.filter(obj => obj.product.optionValue === u?.productDetail?._id).length
+                            tempProducts.find(obj => obj.id === d?.productDetail?._id).qty = u?.qty * d?.qty + tempProducts.find(obj => obj.id === d?.productDetail?._id).qty - alreadyStoredSelectedRecords.filter(obj => obj.product.optionValue === u?.productDetail?._id).length
                         }
                         else {
-                            tempProducts.push({ "id": u?.productDetail?._id, "name": u?.productDetail?.productName || "", "qty": u?.qty * d?.qty - selectedRecords.filter(obj => obj.product.optionValue === u?.productDetail?._id).length })
+                            tempProducts.push({ "id": u?.productDetail?._id, "name": u?.productDetail?.productName || "", "qty": u?.qty * d?.qty - alreadyStoredSelectedRecords.filter(obj => obj.product.optionValue === u?.productDetail?._id).length })
                         }
                     })
                 }
-                if (d?.qty - selectedRecords.filter(obj => obj.product.optionValue === d._id).length <= -1) {
-                    let tempSelectedRecoeds = selectedRecords
-                    var idx = tempSelectedRecoeds.findIndex(obj => obj.product.optionValue === d._id);
-                    var removed = tempSelectedRecoeds.splice(idx, 1);
-                    // dispatch({ type: "loading", loading: true });
-                    // setTimeout(() => {
-                    //     dispatch({ type: "loading", loading: false });
-                    // }, gridLoadingTimeout);
-                    dispatch({ type: "selection", selectedRecords: tempSelectedRecoeds });
-
-                }
+                // if (d?.qty - alreadyStoredSelectedRecords.filter(obj => obj.product.optionValue === d._id).length <= -1) {
+                //     let tempSelectedRecoeds = [...alreadyStoredSelectedRecords]
+                //     var idx = tempSelectedRecoeds.findIndex(obj => obj.product.optionValue === d._id);
+                //     var removed = tempSelectedRecoeds.splice(idx, 1);
+                //     // dispatch({ type: "loading", loading: true });
+                //     // setTimeout(() => {
+                //     //     dispatch({ type: "loading", loading: false });
+                //     // }, gridLoadingTimeout);
+                //     dispatch({ type: "selection", selectedRecords: tempSelectedRecoeds });
+                // }
             })
         }
         setSerializedProducts(tempProducts)
@@ -108,10 +111,16 @@ const AddSerializedAsset = ({ isAdding, addSerializedAsset, handleSerializedAsse
             gridApi.setRowData([]);
         }
 
-        const queryString = getQueryString();
+        let queryString = getQueryString();
+        if (selectedProducts.length > 0) {
+            queryString = `${queryString}&filterById=${JSON.stringify(selectedProducts.map(m => { return { "field": "product", "term": m?._id ?? "" } }))}&filterByIdType=or`
+        }
+
         axiosInstance().get(`${productInventory.api}${queryString}`).then(({ data }) => {
-            data.data = data.data?.filter(u => (u?.status === "Available" || u?.status === "New")
-                && selectedProducts.some(d => d._id === u?.product?.optionValue || d.products?.some(obj => obj?.productId === u?.product?.optionValue)))
+            data.data = data.data
+                // ?.filter(u => (u?.status === "Available" || u?.status === "New")
+                // && selectedProducts.some(d => d._id === u?.product?.optionValue || d.products?.some(obj => obj?.productId === u?.product?.optionValue))
+                // )
                 .map((u) => ({
                     ...u,
                     id: u._id,
@@ -132,10 +141,8 @@ const AddSerializedAsset = ({ isAdding, addSerializedAsset, handleSerializedAsse
         });
     };
 
-
-
     const getQueryString = () => {
-        let deepFilter = `?page=${page}&limit=${limit}`;
+        let deepFilter = `?page=${page}&limit=${limit}&availableAssets=true`;
 
         if (!isObjectEmpty(filters)) {
             const updatedFilters = [];
@@ -213,8 +220,19 @@ const AddSerializedAsset = ({ isAdding, addSerializedAsset, handleSerializedAsse
                     <Box mb={2}>
                         <Grid container >
                             <Grid item xs={12} sm={6}>
-                                {serializedProducts.length > 0 ? serializedProducts.map(d => <span>{d.name ? `  ${d.name} (${d?.qty})  |` : ""}
-                                </span>) : null}
+                                <div>
+                                    {
+                                        serializedProducts.length > 0 ?
+                                            serializedProducts.map(d =>
+                                                d?.qty < 0 ? <span className="text-error">{d.name ? `  ${d.name} (${d?.qty})  |` : ""}</span>
+                                                    : (d?.qty === 0 ? <span className="text-success">{d.name ? `  ${d.name} (${d?.qty})  |` : ""}</span> : <span>{d.name ? `  ${d.name} (${d?.qty})  |` : ""}</span>)
+                                            ) : null
+                                    }
+                                </div>
+
+                                {
+                                    serializedProducts.length > 0 && serializedProducts.some(s => s.qty < 0) ? <div className="text-error font-weight-bold">You have selected more assets then needed.</div> : ""
+                                }
                             </Grid>
                             <Grid item xs={12} sm={6} container justify="flex-end">
                                 <SearchBox
@@ -226,12 +244,13 @@ const AddSerializedAsset = ({ isAdding, addSerializedAsset, handleSerializedAsse
                                 <Box ml={1} mt={1} >
                                     <Button size="small"
                                         color="primary"
-                                        onClick={() => addSerializedAsset(selectedRecords)}
+                                        onClick={() => addSerializedAsset([...getLocalStorageArrayData(localStorageSelectedRecords)])}
                                         variant="contained"
-                                        disabled={selectedRecords.length === 0 || isAdding}
+                                        disabled={getLocalStorageArrayData(`${addSerializedAssetsRenderedFrom}_selected`).length === 0 || isAdding ||
+                                            serializedProducts.some(d => d?.qty < 0)}
                                         endIcon={isAdding && <CircularProgress size={20} />}
                                     >
-                                        {selectedRecords.length ? "(" + selectedRecords.length + ")  " : ""}
+                                        {getLocalStorageArrayData(`${addSerializedAssetsRenderedFrom}_selected`).length ? "(" + getLocalStorageArrayData(`${addSerializedAssetsRenderedFrom}_selected`).length + ")  " : ""}
                                         Add</Button>
                                 </Box>
                             </Grid>
@@ -251,7 +270,10 @@ const AddSerializedAsset = ({ isAdding, addSerializedAsset, handleSerializedAsse
                             allowAction={false}
                             loading={loading}
                             customGridOptions={{ getRowStyle: getRowStyleScheduled }}
-                            selectedRecords={selectedRecords}
+                            // selectedRecords={selectedRecords}
+                            renderedFrom={addSerializedAssetsRenderedFrom}
+                            // showOnlyShowFilteredRecordSwitch={true}
+                            // allowHeaderSelection={false}
                         />
                         : <Box p={2} height={500} bgcolor="white"><CommonSkeleton lenArray={[...Array(10).keys()]} /></Box>}
                 </div>