+++ src/pages/RentalManagement/BulkEditInventoryDialog.tsx
@@ -16,12 +16,12 @@ import {
 import { KeyboardDatePicker, MuiPickersUtilsProvider } from '@material-ui/pickers';
 import DateUtils from '@date-io/date-fns';
 import moment from 'moment';
-
 import { dateFormatForInputControl } from '../../constants/helpers';
 import CustomDialogContent from '../../components/CustomDialog/CustomDialogContent';
 import CustomDialogFooter from '../../components/CustomDialog/CustomDialogFooter';
 import CustomDialogHeader from '../../components/CustomDialog/CustomDialogHeader';
 import { startCase } from 'lodash';
+import axiosInstance from "../../axios/axiosInstance";
 
 interface EditDialogProps {
   onClose: VoidFunction | any;
@@ -35,28 +35,37 @@ interface EditDialogProps {
 }
 
 const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose, isSaving, submitBulkEdit, currencySymbol, data, startDate, endDate }) => {
+
   const [values, setValues] = useState(null);
   const [isDisabled, setDisabled] = useState(false);
   const [errors, setErrors] = useState(null);
-
+  const [units, setUnits] = useState([
+    { "optionLabel": "Price Per Well", "optionValue": "Price Per Well" },
+    { "optionLabel": "Two Well Pad", "optionValue": "Two Well Pad" },
+    { "optionLabel": "Three Well Pad", "optionValue": "Three Well Pad" },
+    { "optionLabel": "Four Well Pad", "optionValue": "Four Well Pad" },
+    { "optionLabel": "Five Well Pad", "optionValue": "Five Well Pad" },
+    { "optionLabel": "Six Well Pad", "optionValue": "Six Well Pad" },
+  ]);
 
   useEffect(() => {
     if (data) {
       if (data.hasOwnProperty('packageId')) {
         setDisabled(true);
       }
-
       const newValues = {
         ...data,
         qty: data?.qty || 0,
         pricingMethod: data?.pricingMethod || 'perDay',
         startDate: data?.startDate || new Date(),
         endDate: data?.endDate || new Date(),
+        tenure: data?.tenure || moment(data?.endDate).diff(moment(data?.startDate), 'days'),
         UOM: data?.UOM || '',
         price: data?.price || 0,
         discount: data?.discount || 0,
         finalPrice: data?.finalPrice || 0
       };
+      newValues["amount"] = data?.amount || newValues.qty * newValues.tenure * newValues.price
       setValues(newValues);
     }
   }, [data]);
@@ -68,10 +77,8 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
   const handleChange = (name: string, value: any) => {
     setValues((prevState) => {
       const newValues = { ...prevState, [name]: value };
-
       return newValues;
     });
-
     let errs = { ...errors };
     if (Boolean(errs?.qty) && name === 'qty' && value) {
       delete errs.qty;
@@ -85,7 +92,6 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
     if (Boolean(errs?.UOM) && name === 'UOM' && value) {
       delete errs.UOM;
     }
-
     if (Object.keys(errs).length === 0) {
       setErrors(null);
     } else {
@@ -98,26 +104,15 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
       if (values?.qty > 0 && values?.pricingMethod !== '' && values?.UOM !== '') {
         const priceData = await calculatePrice([values]);
         if (priceData && priceData.length) {
-          let price = priceData[0].mrp;
-          let qty = priceData[0].qty;
-
+          let price: any = priceData[0].mrp;
           handleChange('price', price);
-
-          const startDate = moment(values?.startDate);
-          const endDate = moment(values?.endDate);
-          const diff = endDate.diff(startDate, 'days');
-
-          let finalPrice = qty && price ? (values?.pricingMethod === 'perDay' && diff !== 0 ? qty * price * diff : qty * price) : price;
-
-          handleChange('finalPrice', finalPrice <= 0 ? 0 : finalPrice);
+          handlePriceCalculation(price, 'price');
         }
       }
     }
   };
 
-  /**
-   * HANDLE CLOSE DIALOG
-   */
+
   const handleClose = () => {
     if (isSaving === false) {
       onClose();
@@ -164,38 +159,101 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
 
   // const getErrorMsg = (str: string) => `${str} is a required field`;
 
-  const finalPriceCalculation = (data: number, type: string) => {
-    const startDate = moment(values?.startDate);
-    const endDate = moment(values?.endDate);
-    const diff = endDate.diff(startDate, 'days');
-    let finalPrice = 0;
-    let discountPrice = values?.finalPrice !== 0 && (values?.finalPrice / 100) * (type === "discount" ? data : values?.discount);
-    let taxPrice = values?.finalPrice !== 0 && (values?.finalPrice / 100) * (type === "tax" ? data : values?.tax);
+  // const finalPriceCalculation = (data: number, type: string) => {
+  //   const startDate = moment(values?.startDate);
+  //   const endDate = moment(values?.endDate);
+  //   const diff = endDate.diff(startDate, 'days');
+  //   let finalPrice = 0;
+  //   let discountPrice = values?.finalPrice !== 0 && (values?.finalPrice / 100) * (type === "discount" ? data : values?.discount);
+  //   let taxPrice = values?.finalPrice !== 0 && (values?.finalPrice / 100) * (type === "tax" ? data : values?.tax);
 
 
-    if (type === 'qty') {
-      finalPrice = data > 0 && values?.price > 0 ? values?.price * data : values?.price
+  //   if (type === 'qty') {
+  //     finalPrice = data > 0 && values?.price > 0 ? values?.price * data : values?.price
+  //   }
+  //   if (type === 'price') {
+  //     finalPrice = values?.qty > 0 && data > 0 ? values?.qty * data : values?.price;
+  //   }
+  //   if (type === 'discount') {
+  //     finalPrice = values?.qty > 0 && values?.price > 0 ? values?.qty * values?.price : values?.price
+  //   }
+  //   if (type === 'tax') {
+  //     finalPrice = values?.qty > 0 && values?.price > 0 ? values?.qty * values?.price : values?.price;
+  //   }
+
+  //   finalPrice = values?.pricingMethod === 'perDay' ? finalPrice * 1 : finalPrice;
+  //   finalPrice = diff > 0 ? finalPrice * diff : finalPrice;
+  //   finalPrice = discountPrice && discountPrice > 0 ? finalPrice - Math.round(discountPrice) : finalPrice;
+
+  //   finalPrice = taxPrice && taxPrice > 0 ? finalPrice + Math.round(taxPrice) : finalPrice;
+
+  //   handleChange('totalTax', taxPrice > 0 ? parseFloat(Math.round(taxPrice).toFixed(1)) : 0);
+  //   handleChange('discountedPrice', discountPrice > 0 ? parseFloat(Math.round(discountPrice).toFixed(1)) : 0);
+  //   handleChange('finalPrice', finalPrice > 0 ? parseFloat(Math.round(finalPrice).toFixed(1)) : 0);
+  // };
+
+
+  const handlePriceCalculation = (value: any, type: string) => {
+
+    const pricingMethod = type === "pricingMethod" ? value : values?.pricingMethod;
+
+    let tenureType: any = "days";
+    if (pricingMethod === "perWeek") {
+      tenureType = "weeks"
     }
-    if (type === 'price') {
-      finalPrice = values?.qty > 0 && data > 0 ? values?.qty * data : values?.price;
-    }
-    if (type === 'discount') {
-      finalPrice = values?.qty > 0 && values?.price > 0 ? values?.qty * values?.price : values?.price
-    }
-    if (type === 'tax') {
-      finalPrice = values?.qty > 0 && values?.price > 0 ? values?.qty * values?.price : values?.price;
+    else if (pricingMethod === "perMonth") {
+      tenureType = "months"
     }
 
-    finalPrice = values?.pricingMethod === 'perDay' ? finalPrice * 1 : finalPrice;
-    finalPrice = diff > 0 ? finalPrice * diff : finalPrice;
-    finalPrice = discountPrice && discountPrice > 0 ? finalPrice - Math.round(discountPrice) : finalPrice;
+    const startDate = moment(type === "startDate" ? value : values?.startDate);
+    let endDate = moment(type === "endDate" ? value : values?.endDate);
 
-    finalPrice = taxPrice && taxPrice > 0 ? finalPrice + Math.round(taxPrice) : finalPrice;
+    if (type === "tenure") {
+      endDate = startDate.add(value, tenureType)
+      handleChange('endDate', endDate);
+    }
 
-    handleChange('totalTax', taxPrice > 0 ? parseInt(Math.round(taxPrice).toFixed(1)) : 0);
-    handleChange('discountedPrice', discountPrice > 0 ? parseInt(Math.round(discountPrice).toFixed(1)) : 0);
-    handleChange('finalPrice', finalPrice > 0 ? parseInt(Math.round(finalPrice).toFixed(1)) : 0);
-  };
+    const tenure = type === "tenure" ? value : endDate.diff(startDate, tenureType) > 0 ? endDate.diff(startDate, tenureType) : 1;
+    handleChange('tenure', tenure);
+
+    const qty = type === "qty" ? value : values?.qty ? values?.qty : 0
+    const price = type === "price" ? value : values?.price ? values?.price : 0
+    let amount = qty * price * tenure
+
+    handleChange('amount', amount);
+
+    let discount = type === "discount" ? value : values?.discount ? values?.discount : 0
+    let discountedPrice = type === "discountedPrice" ? value : values?.discountedPrice ? values?.discountedPrice : 0
+    if (type === "discountedPrice") {
+      discount = parseFloat((100 * discountedPrice / amount).toFixed(2));
+      handleChange('discount', discount);
+    }
+    discountedPrice = parseFloat((amount * discount / 100).toFixed(2));
+    handleChange('discountedPrice', discountedPrice);
+    amount = amount - discountedPrice;
+
+    let tax = type === "tax" ? value : values?.tax ? values?.tax : 0
+    let totalTax = type === "totalTax" ? value : values?.totalTax ? values?.totalTax : 0
+    if (type === "totalTax") {
+      tax = parseFloat((100 * totalTax / amount).toFixed(2));;
+      handleChange('tax', tax);
+    }
+    totalTax = parseFloat((amount * tax / 100).toFixed(2));
+    handleChange('totalTax', totalTax);
+    amount = amount + totalTax;
+
+    handleChange('finalPrice', parseFloat(amount.toFixed(2)));
+  }
+
+
+  useEffect(() => {
+    axiosInstance().get(`/field?resource=Product`).then(({ data: { data } }) => {
+      const unitField = data.filter((e) => e.fieldData.fieldName.toLowerCase().includes("unit") && e.fieldData.type === "dropDown");
+      if (unitField.length) {
+        setUnits(unitField[0].fieldData.option);
+      }
+    })
+  }, []);
 
   return (
     <Dialog open fullWidth maxWidth="md" onClose={handleClose}>
@@ -217,8 +275,8 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
                   variant={'outlined'}
                   value={values?.qty || 0}
                   onChange={(e) => {
-                    let qty = parseInt(e.target.value);
-                    finalPriceCalculation(qty, 'qty');
+                    let qty = parseFloat(e.target.value);
+                    handlePriceCalculation(qty, 'qty');
                     handleChange(e.target.name, qty <= 0 ? 1 : qty);
                   }}
                 />
@@ -237,7 +295,10 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
                     labelId="pricing-method-label"
                     label="Pricing Method"
                     value={values?.pricingMethod || ''}
-                    onChange={(e) => handleChange(e.target.name, e.target.value)}
+                    onChange={(e) => {
+                      handlePriceCalculation(e.target.value, 'pricingMethod');
+                      handleChange(e.target.name, e.target.value)
+                    }}
                   >
                     <MenuItem value="perDay">Per Day</MenuItem>
                     <MenuItem value="perWeek">Per Week</MenuItem>
@@ -259,9 +320,28 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
                   format={dateFormatForInputControl}
                   autoOk
                   value={values?.startDate || new Date(startDate)}
-                  onChange={(date) => handleChange('startDate', date)}
+                  onChange={(date) => {
+                    handlePriceCalculation(date, 'startDate');
+                    handleChange('startDate', date)
+                  }}
                   InputLabelProps={{
                     shrink: true
+                  }}
+                />
+              </Grid>
+              <Grid item xs={12} sm={6}>
+                <TextField
+                  label={"Expected No. of " + (values?.pricingMethod ? values.pricingMethod.replace("per", "") : "Tenure")}
+                  name="tenure"
+                  fullWidth
+                  size="small"
+                  type="number"
+                  variant={'outlined'}
+                  value={values?.tenure || 0}
+                  onChange={(e) => {
+                    let tenure = parseInt(e.target.value);
+                    handlePriceCalculation(tenure, 'tenure');
+                    handleChange(e.target.name, tenure <= 0 ? 0 : tenure);
                   }}
                 />
               </Grid>
@@ -278,7 +358,10 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
                   value={values?.endDate || new Date(endDate)}
                   minDate={new Date(startDate)}
                   maxDate={new Date(endDate)}
-                  onChange={(date) => handleChange('endDate', date)}
+                  onChange={(date) => {
+                    handlePriceCalculation(date, 'endDate');
+                    handleChange('endDate', date)
+                  }}
                   InputLabelProps={{
                     shrink: true
                   }}
@@ -300,7 +383,9 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
                     value={values?.UOM || ''}
                     onChange={(e) => handleChange(e.target.name, e.target.value)}
                   >
-                    <MenuItem value="pcs">Pcs</MenuItem>
+                    {units && units.map((_u: any) => {
+                      return <MenuItem value={_u.optionValue}>{_u.optionLabel}</MenuItem>
+                    })}
                   </Select>
                   {/* <FormHelperText id="UOM-label">{Boolean(errors?.UOM) && errors.UOM}</FormHelperText> */}
                 </FormControl>
@@ -320,10 +405,25 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
                   variant={'outlined'}
                   value={values?.price || 0}
                   onChange={(e) => {
-                    let price = parseInt(e.target.value);
-                    finalPriceCalculation(price, 'price');
+                    let price = parseFloat(e.target.value);
+                    handlePriceCalculation(price, 'price');
                     handleChange(e.target.name, price <= 0 ? 0 : price);
                   }}
+                />
+              </Grid>
+              <Grid item xs={12} sm={6}>
+                <TextField
+                  label="Total Price"
+                  name="amount"
+                  fullWidth
+                  InputProps={{
+                    startAdornment: <InputAdornment position="start">{currencySymbol ? currencySymbol : ''}</InputAdornment>,
+                    readOnly: true
+                  }}
+                  size="small"
+                  type="number"
+                  variant={'outlined'}
+                  value={values?.amount || 0}
                 />
               </Grid>
               <Grid item xs={12} sm={6}>
@@ -340,15 +440,15 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
                   variant={'outlined'}
                   value={values?.discount || 0}
                   onChange={(e) => {
-                    let discount = parseInt(e.target.value) && parseInt(e.target.value) > 0 ? parseInt(e.target.value) : 0;
-                    finalPriceCalculation(discount, 'discount');
+                    let discount = parseFloat(e.target.value) && parseFloat(e.target.value) > 0 ? parseFloat(e.target.value) : 0;
+                    handlePriceCalculation(discount, 'discount');
                     handleChange(e.target.name, discount);
                   }}
                 />
               </Grid>
               <Grid item xs={12} sm={6}>
                 <TextField
-                  label="Discounted Price"
+                  label="Discount"
                   name="discountedPrice"
                   fullWidth
                   InputProps={{
@@ -359,7 +459,11 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
                   type="number"
                   variant={'outlined'}
                   value={values?.discountedPrice || 0}
-                  onChange={(e) => handleChange(e.target.name, parseInt(e.target.value) <= 0 ? 0 : parseInt(e.target.value))}
+                  onChange={(e) => {
+                    let discountedPrice = parseFloat(e.target.value) <= 0 ? 0 : parseFloat(e.target.value);
+                    handlePriceCalculation(discountedPrice, 'discountedPrice');
+                    handleChange(e.target.name, discountedPrice)
+                  }}
                 />
               </Grid>
               <Grid item xs={12} sm={6}>
@@ -376,15 +480,15 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
                   variant={'outlined'}
                   value={values?.tax || 0}
                   onChange={(e) => {
-                    const tax = parseInt(e.target.value) && parseInt(e.target.value) > 0 ? parseInt(e.target.value) : 0
-                    finalPriceCalculation(tax, "tax")
+                    const tax = parseFloat(e.target.value) && parseFloat(e.target.value) > 0 ? parseFloat(e.target.value) : 0
+                    handlePriceCalculation(tax, "tax")
                     handleChange(e.target.name, tax)
                   }}
                 />
               </Grid>
               <Grid item xs={12} sm={6}>
                 <TextField
-                  label="Total Tax"
+                  label="Tax"
                   name="totalTax"
                   fullWidth
                   InputProps={{
@@ -395,7 +499,11 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
                   type="number"
                   variant={'outlined'}
                   value={values?.totalTax || 0}
-                  onChange={(e) => handleChange(e.target.name, parseInt(e.target.value) <= 0 ? 0 : parseInt(e.target.value))}
+                  onChange={(e) => {
+                    const totalTax = parseFloat(e.target.value) <= 0 ? 0 : parseFloat(e.target.value)
+                    handlePriceCalculation(totalTax, "totalTax")
+                    handleChange(e.target.name, totalTax)
+                  }}
                 />
               </Grid>
               <Grid item xs={12} sm={6}>
@@ -411,7 +519,7 @@ const BulkEditInventoryDialog: FC<EditDialogProps> = ({ calculatePrice, onClose,
                   type="number"
                   variant={'outlined'}
                   value={values?.finalPrice || 0}
-                  onChange={(e) => handleChange(e.target.name, parseInt(e.target.value) <= 0 ? 0 : parseInt(e.target.value))}
+                  onChange={(e) => handleChange(e.target.name, parseFloat(e.target.value) <= 0 ? 0 : parseFloat(e.target.value))}
                 />
               </Grid>
             </Grid>